Move const to tailwind config package
